Move UI components to appropriate sub-packages
