+++ noty-android/app/composeapp/src/main/java/dev/shreyaspatil/noty/composeapp/ui/theme/Shape.kt
@@ -14,18 +14,14 @@
  * limitations under the License.
  */
 
-package dev.shreyaspatil.noty.composeapp.view
+package dev.shreyaspatil.noty.composeapp.ui.theme
 
-sealed class Screen(val route: String, val name: String) {
-    object SignUp : Screen("signup", "Sign Up")
-    object Login : Screen("login", "Login")
-    object Notes : Screen("notes", "Notes")
-    object NotesDetail : Screen("note/{noteId}", "Note details") {
-        fun route(noteId: String) = "note/$noteId"
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Shapes
+import androidx.compose.ui.unit.dp
 
-        const val ARG_NOTE_ID: String = "noteId"
-    }
-
-    object AddNote : Screen("note/new", "New note")
-    object About : Screen("about", "About")
-}
+val shapes = Shapes(
+    small = RoundedCornerShape(4.dp),
+    medium = RoundedCornerShape(4.dp),
+    large = RoundedCornerShape(0.dp)
+)
