+++ noty-android/app/src/main/java/dev/shreyaspatil/noty/view/viewmodel/NoteDetailViewModel.kt
@@ -16,6 +16,7 @@
 
 package dev.shreyaspatil.noty.view.viewmodel
 
+import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
@@ -117,6 +118,11 @@ class NoteDetailViewModel @AssistedInject constructor(
 
     private fun scheduleNoteDelete(noteId: String) =
         notyTaskManager.scheduleTask(NotyTask.delete(noteId))
+
+    override fun onCleared() {
+        super.onCleared()
+        Log.d("NoteDetailViewModel", "onCleared")
+    }
 
     @AssistedFactory
     interface Factory {
